Clustering added (#49)
* Clustering added

* Rewritten with prettier
