+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 node_js:
-  - "15"
+  - '15'
 script:
   - echo 'Build starts!!'
   - echo 'Installing Deps!'
