+++ config.js
@@ -1,6 +1,6 @@
-import { config } from "dotenv";
+import dotenv from 'dotenv';
 
-config();
+dotenv.config();
 
 export const db_host = process.env.db_host;
 export const db_name = process.env.db_name;
