+++ documentation/ideas.md
@@ -1,1 +1,1 @@
-Document New Features/Ideas here 
+Document New Features/Ideas here
