+++ package.json
@@ -21,6 +21,7 @@
   "homepage": "https://github.com/Manvityagi/Opportunity-Calendar-Backend#readme",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "cluster": "^0.7.7",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
