+++ routes/index.js
@@ -1,7 +1,6 @@
 import express from 'express';
 const router = express();
 
-
 /**
  * @openapi
  * /:
